fix(ui): remove airflow from generic form component
